% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{clust}
\alias{clust}
\title{画出聚类图}
\usage{
clust(expr_mat, clust_method = "complete")
}
\arguments{
\item{expr_mat}{表达矩阵}

\item{clust_method}{聚类方法，可选 "complete", "single", "average", "ward.D"等}
}
\description{
画出聚类图
}
