% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOKEGG.R
\name{enrich_split}
\alias{enrich_split}
\title{对up和down分别进行GO/KEGG分析}
\usage{
enrich_split(up, down, method = "GO", pvaluecutoff = 0.05, n = 5)
}
\arguments{
\item{up}{带EntrezID列的上调基因}

\item{down}{带EntrezID列的下调基因}

\item{method}{"GO" / "KEGG"}

\item{pvaluecutoff}{pvalue阈值}

\item{n}{图中显示项的数量}
}
\description{
对up和down分别进行GO/KEGG分析
}
