% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{getModuleGenes}
\alias{getModuleGenes}
\title{获取指定模块内的基因}
\usage{
getModuleGenes(expr_mat, result, module)
}
\arguments{
\item{expr_mat}{表达矩阵}

\item{result}{WGCNA分析结果}

\item{module}{模块颜色}
}
\description{
获取指定模块内的基因
}
