% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{limmaAnalysis}
\alias{limmaAnalysis}
\title{limma差异分析}
\usage{
limmaAnalysis(
  expr_mat,
  traits_mat,
  contrasts,
  item = 1,
  pcutoff = 0.05,
  lfc = 0,
  title = "Volcano Plot"
)
}
\arguments{
\item{expr_mat}{表达矩阵}

\item{traits_mat}{性状矩阵}

\item{contrasts}{对比项}

\item{item}{获取哪个对比项的差异分析结果}

\item{pcutoff}{pvalue阈值}

\item{lfc}{log foldchange阈值}

\item{title}{火山图标题}
}
\description{
limma差异分析
}
