% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{matchEntrezID}
\alias{matchEntrezID}
\title{根据基因名匹配EntrezID}
\usage{
matchEntrezID(DA_result, anno)
}
\arguments{
\item{DA_result}{limmaAnalysis返回的结果}

\item{anno}{标注信息，包含Symbol和ENTREZID列}
}
\description{
根据基因名匹配EntrezID
}
