% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{removeDuplication}
\alias{removeDuplication}
\title{去除表达矩阵中重复的基因}
\usage{
removeDuplication(dat)
}
\arguments{
\item{dat}{最后两列为Symbol和ENTREZID的表达矩阵}
}
\value{
去重并且以Symbol为行名的表达矩阵
}
\description{
保留表达值平均数最大的基因
}
