% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{WGCNAHeatmap}
\alias{WGCNAHeatmap}
\title{Title}
\usage{
WGCNAHeatmap(WGCNA_result, traits_mat)
}
\arguments{
\item{WGCNA_result}{WGCNA分析结果}

\item{traits_mat}{性状矩阵}
}
\description{
画出module与性状的相关性热图
}
