% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{WGCNAMergeModule}
\alias{WGCNAMergeModule}
\title{Title}
\usage{
WGCNAMergeModule(expr_mat, WGCNA_result, cutHeightThres = 0.4)
}
\arguments{
\item{expr_mat}{表达矩阵}

\item{WGCNA_result}{WGCNA分析结果}

\item{cutHeightThres}{cut阈值}
}
\description{
根据cut高度合并module
}
