% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BendchipAnalysis.R
\name{WGCNAScatterPlot}
\alias{WGCNAScatterPlot}
\title{画WGCNA分析的散点图}
\usage{
WGCNAScatterPlot(result, traits_mat, module, pheno)
}
\arguments{
\item{result}{WGCNA分析结果}

\item{traits_mat}{性状矩阵}

\item{module}{模块颜色}

\item{pheno}{性状名}
}
\description{
画WGCNA分析的散点图
}
